requirements.txt
garminconnect
python-dotenv
pandas
plotly
nbformat
ipykernel